+++ cmd/internal/problem_test.go
@@ -10,10 +10,10 @@ func TestNewProblem(t *testing.T) {
 	config, _ := NewConfig("fixtures/002_020_slice.json")
 	problem, _ := NewProblem(config)
 
-	assert.Equal(problem.cc, uint32(2), t)
-	assert.Equal(problem.tc, uint32(20), t)
-	assert.Equal(problem.uc, uint32(20), t)
-	assert.Equal(problem.zc, uint32(3), t)
+	assert.Equal(problem.cc, uint(2), t)
+	assert.Equal(problem.tc, uint(20), t)
+	assert.Equal(problem.uc, uint(20), t)
+	assert.Equal(problem.zc, uint(3), t)
 
 	delay := make([]float64, 20)
 	for i := 0; i < 20; i++ {
@@ -27,11 +27,11 @@ func TestNewProblem(t *testing.T) {
 
 	assert.Equal(len(problem.multiplier), 3*20, t)
 
-	assert.Equal(problem.schedule.Mapping, []uint16{
+	assert.Equal(problem.schedule.Mapping, []uint{
 		0, 1, 0, 0, 1, 1, 1, 0, 0, 1,
 		1, 0, 0, 0, 0, 1, 1, 1, 1, 1,
 	}, t)
-	assert.Equal(problem.schedule.Order, []uint16{
+	assert.Equal(problem.schedule.Order, []uint{
 		0, 1, 2, 9, 12, 16, 18, 14, 17, 13,
 		15, 3, 5, 11, 19, 8, 7, 6, 4, 10,
 	}, t)
